feat: custom-spending-row 적용
