+++ pennyway-client-iOS/pennyway-client-iOS/View/TabView/SpendingManagementView/MySpendingListView/MySpendingListView.swift
@@ -40,7 +40,7 @@ struct MySpendingListView: View {
                                             ForEach(spendings, id: \.id) { item in
                                                 let iconName = SpendingListViewCategoryIconList(rawValue: item.category.icon)?.iconName ?? ""
                                                 NavigationLink(destination: DetailSpendingView()) {
-                                                    ExpenseRow(categoryIcon: iconName, category: item.category.name, amount: item.amount, memo: item.memo)
+                                                    CustomSpendingRow(categoryIcon: iconName, category: item.category.name, amount: item.amount, memo: item.memo)
                                                 }
 
                                                 Spacer().frame(height: 12 * DynamicSizeFactor.factor())
@@ -201,51 +201,6 @@ struct MySpendingListView: View {
             return dateFormatter.string(from: previousMonthDate)
         }
         return dateFormatter.string(from: date)
-    }
-}
-
-// MARK: - ExpenseRow
-
-struct ExpenseRow: View {
-    var categoryIcon: String
-    var category: String
-    var amount: Int
-    var memo: String
-
-    var body: some View {
-        ZStack(alignment: .leading) {
-            HStack(spacing: 10 * DynamicSizeFactor.factor()) {
-                Image(categoryIcon)
-                    .resizable()
-                    .frame(width: 40 * DynamicSizeFactor.factor(), height: 40 * DynamicSizeFactor.factor())
-
-                VStack(alignment: .leading, spacing: 1) { // Spacer는 Line heigth 적용하면 없애기
-                    if memo.isEmpty {
-                        Text(category)
-                            .font(.B1SemiboldeFont())
-                            .platformTextColor(color: Color("Gray06"))
-                            .multilineTextAlignment(.leading)
-                    } else {
-                        Text(category)
-                            .font(.B1SemiboldeFont())
-                            .platformTextColor(color: Color("Gray06"))
-                            .multilineTextAlignment(.leading)
-
-                        Text(memo)
-                            .font(.B3MediumFont())
-                            .platformTextColor(color: Color("Gray04"))
-                            .multilineTextAlignment(.leading)
-                    }
-                }
-
-                Spacer()
-
-                Text("\(amount)원")
-                    .font(.B1SemiboldeFont())
-                    .platformTextColor(color: Color("Gray06"))
-            }
-        }
-        .padding(.horizontal, 20)
     }
 }
